Make script idempotent

+++ tasks/run-local-gats/run.ps1
@@ -108,7 +108,9 @@ push-location garden-runc-release
   # Upstream garden-runc-release switched to go-modules
   # We still use GOPATH, so move guardian accordingly
   mkdir ./src/gopath/src/code.cloudfoundry.org -ea 0
-  mv ./src/guardian ./src/gopath/src/code.cloudfoundry.org/
+  if (Test-Path ./src/guardian) {
+    mv ./src/guardian ./src/gopath/src/code.cloudfoundry.org/
+  }
   go build -o gdn.exe ./src/gopath/src/code.cloudfoundry.org/guardian/cmd/gdn
   if ($LastExitCode -ne 0) {
       throw "Building gdn.exe process returned error code: $LastExitCode"
